Only try to build ContikiMote Java library once.
Cooja tries to build the ContikiMote Java library first without and then with
the Cooja JAR included in the class path. This results in a warning if it
fails in the first try. This changes Cooja to try to compile the library both
with current path and the Cooja JAR in the class path in a single step to
avoid the compile warning.a single step.