Fix clock deviation in MspMote
We measured the impact of the deviation parameter on the mote execution
and found significant inaccuracies. These would only appear when LPM was
enabled on the mote.

We identified the origin of these inaccuracies to incorrect estimation
of the timeline increase in COOJA/MSPSIM when both clock deviation and
LPM are involved.

This new implementation of the execution for MspMote can precisely
reproduce the requested deviation with absolute clock drift error well
below 1 ppm.

More details about this evaluation and the new implementation can be
found in:
 "Toward accurate clock drift modeling in Wireless Sensor Networks
 simulation"
 https://doi.org/10.1016/j.comcom.2020.08.025