Update to log4j 2.17.2
Log4j 1.x reached end of life in 2015 so
update to the latest Log4j 2.x.

This commit uses the API bridge to Log4j 1.x
to minimize the changes, but since the import
statement and how to get the logger changed
we still have to change 160+ files.

Also add final to the Logger while we touch
these lines.