Stop interfering with thread scheduler
The documentation for yield says:

It is rarely appropriate to use this method.
It may be useful for debugging or testing
purposes, where it may help to reproduce
bugs due to race conditions.

Beyond the anti-recommendation from the
documentation, we want to keep the thread
active so replace the yield with a short sleep
instead. Also remove the call to setPriority
and let the simulation thread keep the default
priority.

The net result seems to be a slight performance
improvement, using a precompiled jar and running
time -p make in 07-simulation-base in docker
before the change:

real 181.82
user 356.84
sys 30.90

and after:

real 175.49
user 342.36
sys 28.92