Enable Cooja to pass core count to make
Add a variable $(CPUS) to commands that
is replaced with the number of cores,
so we can avoid spawning 40 compilers with
"make -j" in the Contiki-NG regression tests.

This gives a slight performance improvement
on 07-simulation-tests on my 12 core machine,
time -p make before the change:

real 174.78
user 339.52
sys 27.04

and after the change:

real 169.06
user 330.62
sys 26.70

Machines with fewer cores should see a greater
improvement.