Compile LibN.java in temporary directory
This ensures Cooja will either get a LibN.class
that corresponds to the freshly generated LibN.java,
or fail.

A secondary benefit is that Cooja will no longer
leave the "org" directory in the regression tests
in the Contiki-NG source tree.

This commit also fixes a bug where the return value
of the javac invocation was never checked.