Allow the user to configure log directory
Cooja currently outputs logs to the directory
where it is started. This is inconvenient with
the run targets in build.xml since they change
directory to build, so everything gets rebuilt
on each invocation.

Add a command line switch to configure
the log directory and give it a default
value of the current directory to mimic
the current behavior.

The log directory is currently not respected
by log4j, that will be addressed in the next
commit.