build.xml: use jar for run targets
Stop listing the transitive closure of
dependencies on each target and instead
depend on what it needs. Switch run targets
to use the jar file since they already
depend on jar, and remove the environment
setup.

This enables consolidating the two different
classpaths and removing the init target which
is not doing anything useful.