Make log4j configuration programmatic
Rename the log4j parameter to log4j2 to
indicate that the expectation is a version 2.x
configuration file.

Add a parameter logname that lets the user
decide what the log file should be called.
The default is the same as today, COOJA.log.

Construct a default log4j2 config inside Cooja
that is similar to the previous XML file
configuration, except that it respects the logdir
parameter. ExecuteJAR does not accept parameters,
so use the default log4j configuration for that.

The consequence of this is that the log4j2 api bridge
and the log4j 1.x configuration file can be removed.