Distinguish make target and library file names
The firmware for Contiki motes is called mtypeNNN
where NNN is the mote identifier. The static method
getExpectedFirmwareFile returns something else,
and AbstractCompileDialog checks that this file exists.
The file no longer exists after the latest build system
updates, so make getExpectedFirmwareFile return the right
thing.

The make target name is constructed from the name
of the firmware file so that breaks from this fix.
Add a separate method that works like the old
getExpectedFirmwareFile and use that method for
constructing make target names.