Remove the configuration wizard
The responsibility for building motes has been shifted
from Cooja into Contiki-NG. The Contiki-NG build system
gives error messages so the C compiler tests performed
by the wizard are no longer required.

The Java tests performed by the wizard belong in Cooja,
but the regular code paths for that functionality now
contain error checking and provides error messages.

The remaining test performed by the wizard is the symbol
reading. That test is difficult to perform without
the previous tests, and the regular code path does
provide (cryptic) error messages when that fails. The error
detection and error messages for that could be improved.

Beyond the fact that the wizard currently does not work,
removing it also simplifies revising the interface
to CompileContiki.