Simplify the register methods
The return values from the register methods are not
checked so remove the reflection-based checks and
just add the classes to the containers.

This change only affects plugin developers that forget
to add the right constructor to their class. They
will get an error message when the loading fails
instead.

This fixes an ErrorProne error about not checking
the return value of getConstructor.