Cooja: use first <simulation> in config
The current code will allocate as many simulations
as there are <simulation> tags in the config file,
and then use the last allocated simulator, unless
one of the simulation configs are faulty.

Cooja does not handle multiple simulations in this
fashion, so make the code use a single simulation.
This makes it possible to allocate the simulator
outside the try-block which further simplifies
the code.