Simulation: fix speedup printout
Calculate the runtime of the simulation
once, and then use and use the runtime
for calculating the speedup, rather than
calculating a new runtime for the ratio.
Remove the system time in milliseconds part
of the printout, if time should be printed
for informational purposes it should be
a human-readable readable form.

Also move the printout into the part that
leaves the event loop. This makes the runtime
calculation more precise, and avoids printing
the speedup message when exceptions were thrown.