ScriptRunner: minimize try-block
Put the conditional in the argument to activateScript
since the method is called in both modes and push
the GUI updates after the catch-block. The GUI updates
should not be done on failure so return from the catch
block.

Also remove the debug printout, anyone debugging
the code can add it back.