Cooja: add option for using FlatLaf
Add a command-line option to select look and
feel, but continue to use Nimbus as default.

This allows us to remove the fallback code
for setting look and feel, so the user either
gets what they requested, or a failure.

Setting dragMode leaves residue from the borders
when moving windows on OS X, so use the default
dragMode.