Cooja: load known plugins without reflection
Add a switch-statement in setPluginsConfigXML
that either calls the constructor for a known
plugin, or calls the reflection-based alternative
for other plugins.

This avoids the reflection for plugins on
the fast path for the Contiki-NG regression
tests.

Removing the remaining users of reflection for
known plugins will still require some work, but
this is a step forward.