ContikiMoteType: use FFI for offset calculation
Use the Java 17 FFI to calculate the offset.
The main benefit is using a supported API
for getting the variable address, but the change
also avoids an extra copy of the memory.

This commit makes the JVM print a warning:

WARNING: Using incubator modules: jdk.incubator.foreign

and that warning seems to be impossible to disable.