ScriptRunner: log to separate files
This does not change the logging behavior for
a single ScriptRunnner, as long as the log file
can be opened. If the log file cannot be
opened, this will throw an exception before
starting the simulation.

If there are multiple ScriptRunnners, the second
will log to COOJA-02.testlog, the third to
COOJA-03.testlog, and so on.