LogScriptEngine: shutdown without exceptions
Change the return mechanism of scripts from
throwing exceptions even on successful exits
into being a regular return value that is
inspected upon return.

The implementation still uses exceptions inside
the script, but there is an outer catch that
converts the exception into a return value.

This also eliminates the need for spawning
a thread to exit Cooja in headless mode,
the script thread can just call doQuit after
it is done.