BaseContikiMoteType: add setBaseConfigXML method
Put the common configuration work in the base
class, and call that from the subclasses.

This means that ContikiMoteType will now accept
specifying compile commands with "command" instead
of "commands". The same thing happens for MspMoteType
that now accepts "contikibasedir", and gives an error
message when that is encountered. These elements
are never written by the respective mote types though,
so the leniency should not be a problem in practice.