AbstractCompileDialog: remove compilationEnvironment
Stop holding the state in the dialog and just query
the mote type for the environment on each call to compile.

This required putting the environment tab in the tab pane
before a source file has been selected, but that does not
make anything worse.

Also remove the code that is supposed to select the environment
tab after modifying the environment. The environment tab
is already selected since the button is in that tab, so
the effect of the code was to select the advanced tab instead.