ProjectDirectoriesDialog: use JFileChooser
Keep the logic for identifying and marking
Cooja projects/extensions in the GUI, but
replace the JTree on the left with a JFileChooser
behind an add button instead.

This saves 350 lines of code, and instead
of navigating from the root of the filesystem
and having to expand each level, the dialog
uses the Cooja directory as the starting
directory.