GUI: fix the abort button
Make the abort button in the progress dialog
work. This is accomplished by loading simulation
files in the background thread, but creating
the simulation in the AWT thread, and then
invokeLater to transfer the simulation back
to the background thread.

This is preparation for letting the user keep
the current simulation when loading a new simulation
fails.