Visualizer: work around race condition
The last thing the simulation thread does
is call radioMedium.removed(), and some
radio mediums will call
Visualizer.unregisterVisualizerSkin to
unregister the skin they have registered.
If Cooja is loading a new simulation while
this is happening, the new simulation might
fail to load its visualizer skin because
the old simulation removed the skin midway
through the load of the new simulation.

There is no way to get a handle of the current
visualizer, so just move the registration to
inside the Visualizer constructor to avoid
the issue for now.

Long term, skins should have a way to register
themselves without having to call static
methods in the Visualizer.