Cooja: stop calling System.gc()
Removing this seems to save 1-2 seconds on
07-simulation-base.

Before:
real 78.07
user 1.44
sys 0.18

After:
real 77.23
user 1.32
sys 0.27

Cooja has 2 GB heap by default, so there
should not be a shortage of memory for
typical use. If Cooja still needs to
call System.gc() somewhere, it would
probably be better to do it last in
the simulation thread.