Cooja: avoid initializing class containers
Add a lazy initialization of plugins/motes/etc
that is used in headless mode. This keeps the eager
error checking in graphical mode, and enables
Cooja to run the Contiki-NG tests without
touching these containers at all.

Beyond the performance benefits, this should
enable the code for handling plugins/etc to be moved
to a separate class in the future.