Add an extension manager and use it
This is just a skeleton, but the class is
meant to be the place where all the logic
relating to loading and querying extensions
reside.

Make the plugin registration the first
client, and do that a bit more eagerly
in preparation for ensuring only loaded
plugins are used by the simulation.