ContikiClock: fix 64-bit variable size
The clock_time_t definition in Contiki-NG
has been "unsigned long", which is 32 bits
on most 32-bit architectures, and 64 bits
on most 64-bit architectures.

Assume people are running Cooja on 64-bit
machines for now, and read/write as a 64-bit
variable. There is a pull request for
Contiki-NG to make the Cooja platform's
clock_time_t an uint64_t, so this commit
will eventually become correct for 32-bit
machines as well.